// src/schemas/student/student-list/institute-schema.ts
const SessionSchema = z.object({
    id: z.string(),
    session_name: z.string(),
    status: z.string(),
    start_date: z.string(),
});

export const LevelSchema = z.object({
    id: z.string(),
    level_name: z.string(),
    duration_in_days: z.number().nullable(),
    thumbnail_id: z.string().nullable(),
});

export const PackageSchema = z.object({
    id: z.string(),
    package_name: z.string(),
    thumbnail_id: z.string().nullable().optional(),
});

export const BatchForSessionSchema = z.object({
    id: z.string(),
    level: LevelSchema,
    session: SessionSchema,
    start_time: z.string().nullable(),
    status: z.string(),
    package_dto: PackageSchema,
    is_org_associated: z.boolean().optional(),
    group: z.any().nullable().optional(),
    is_parent: z.boolean().optional(),
    parent_id: z.string().nullable().optional(),
});
// ...
export type BatchForSessionType = z.infer<typeof BatchForSessionSchema>;